image_analyzer_backend/requirements.txt
fastapi
uvicorn[standard]
websockets
python-multipart
pillow
tensorflow # For ML operations (CPU version by default unless specified)
ultralytics # For YOLO models